refactor(ontology): rename RedactionRecord to AuditEntry, flatten provenance
- Rename RedactionRecord → AuditEntry, RedactionRecordBuilder → AuditEntryBuilder
- Add timestamp, correlation_id, status fields to AuditEntry
- Flatten provenance/record/* into provenance/* (remove record/ subdir)
- Merge RedactionLifecycle + old ReviewDecision into new ReviewDecision
  with version, status, reviewer_id, timestamp, reason
- Make review: Option<ReviewDecision> on AuditEntry (entire review optional)
- Remove old AuditEntry, AuditEntryKind, InferenceKind, ProcessingKind,
  LifecycleKind, and all action types (InferenceAction, ProcessingAction,
  LifecycleAction)
- Audit.records → Audit.entries (Vec<AuditEntry>)
- Update all engine references
