+++ crates/nvisy-ontology/src/provenance/evaluation.rs
@@ -4,7 +4,7 @@ use schemars::JsonSchema;
 use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
-use super::redaction::RedactionRecord;
+use super::entry::AuditEntry;
 
 /// Full outcome of evaluating a policy against a set of entities.
 #[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
@@ -12,9 +12,9 @@ use super::redaction::RedactionRecord;
 pub struct PolicyEvaluation {
     /// Identifier of the policy that was evaluated.
     pub policy_id: Uuid,
-    /// Redaction records produced by `Redact` rules.
+    /// Audit entries produced by `Redact` rules.
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
-    pub records: Vec<RedactionRecord>,
+    pub records: Vec<AuditEntry>,
     /// Entity IDs routed to human review by `Review` rules.
     #[serde(default, skip_serializing_if = "Vec::is_empty")]
     pub pending_review: Vec<Uuid>,
