--- crates/nvisy-ontology/src/provenance/record/mod.rs
@@ -1,12 +1,0 @@
-//! Redaction record types: audit records, policy evaluations,
-//! and human review.
-
-mod evaluation;
-mod redaction;
-mod review;
-
-pub use self::evaluation::PolicyEvaluation;
-pub use self::redaction::{
-    RedactionLifecycle, RedactionRecord, RedactionRecordBuilder, RedactionSpec, RedactionValue,
-};
-pub use self::review::{ReviewDecision, ReviewStatus};
