+++ crates/nvisy-ontology/src/provenance/review.rs
@@ -6,7 +6,7 @@ use serde::{Deserialize, Serialize};
 use strum::{Display, EnumString};
 use uuid::Uuid;
 
-/// Status of a human review on a redaction decision.
+/// Status of a human review on a redaction.
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 #[derive(Display, EnumString, Serialize, Deserialize, JsonSchema)]
 #[serde(rename_all = "snake_case")]
@@ -23,10 +23,16 @@ pub enum ReviewStatus {
     AutoApproved,
 }
 
-/// A review decision recorded against a redaction.
+/// A review decision recorded against a redaction, including versioning.
+///
+/// Present on an [`AuditEntry`](super::AuditEntry) only when the
+/// redaction has been reviewed (or is pending review). Absent for
+/// entries that have not entered the review workflow.
 #[derive(Debug, Clone, Serialize, Deserialize, JsonSchema)]
 #[serde(rename_all = "camelCase")]
 pub struct ReviewDecision {
+    /// Version of the audit entry at review time (starts at 1).
+    pub version: u32,
     /// Outcome of the review.
     pub status: ReviewStatus,
     /// Identifier of the reviewer (human or service account).
